Conflicts:
	plugins/vcs-core/core/models/diff/merge.mps
	plugins/vcs/common/models/changesmanager.mps
	plugins/vcs/common/models/changesmanager/editor.mps
	plugins/vcs/common/models/diff/ui.mps
	plugins/vcs/common/models/diff/ui/common.mps
	plugins/vcs/common/models/diff/ui/merge.mps
	plugins/vcs/common/source_gen/jetbrains/mps/vcs/diff/ui/MetadataUtil.java
	plugins/vcs/models/jetbrains/mps/vcs/changesmanager/tree.mps
	plugins/vcs/models/jetbrains/mps/vcs/changesmanager@tests.mps
	plugins/vcs/models/jetbrains/mps/vcs/diff/merge@tests.mps

+++ editor/editor-runtime/source/jetbrains/mps/nodeEditor/cells/EditorCell_Label.java
@@ -21,6 +21,7 @@ import com.intellij.util.ui.UIUtil;
 import jetbrains.mps.editor.runtime.cells.AbstractCellAction;
 import jetbrains.mps.editor.runtime.style.Padding;
 import jetbrains.mps.editor.runtime.style.StyleAttributes;
+import jetbrains.mps.editor.runtime.style.StyleAttributesUtil;
 import jetbrains.mps.ide.datatransfer.CopyPasteUtil;
 import jetbrains.mps.ide.datatransfer.TextPasteUtil;
 import jetbrains.mps.nodeEditor.CellSide;
@@ -30,6 +31,7 @@ import jetbrains.mps.nodeEditor.cellMenu.NodeSubstitutePatternEditor;
 import jetbrains.mps.nodeEditor.selection.EditorCellLabelSelection;
 import jetbrains.mps.nodeEditor.text.TextBuilder;
 import jetbrains.mps.openapi.editor.EditorContext;
+import jetbrains.mps.openapi.editor.cells.CellAction;
 import jetbrains.mps.openapi.editor.cells.CellActionType;
 import jetbrains.mps.openapi.editor.cells.SubstituteInfo;
 import jetbrains.mps.openapi.editor.selection.MultipleSelection;
@@ -152,18 +154,20 @@ public abstract class EditorCell_Label extends EditorCell_Basic implements jetbr
     getStyle().set(StyleAttributes.NULL_SELECTED_TEXT_BACKGROUND_COLOR, color);
   }
 
+  /**
+   * @deprecated Use StyleAttributesUtil.isFirstPositionAllowed() instead. Should be removed after MPS 3.0
+   */
+  @Deprecated
   public boolean isFirstPositionAllowed() {
-    if (getStyle().isSpecified(StyleAttributes.FIRST_POSITION_ALLOWED)) {
-      return getStyle().get(StyleAttributes.FIRST_POSITION_ALLOWED);
-    }
-    return !getStyle().get(StyleAttributes.PUNCTUATION_LEFT);
+    return StyleAttributesUtil.isFirstPositionAllowed(getStyle());
   }
 
+  /**
+   * @deprecated Use StyleAttributesUtil.isLastPositionAllowed() instead. Should be removed after MPS 3.0
+   */
+  @Deprecated
   public boolean isLastPositionAllowed() {
-    if (getStyle().isSpecified(StyleAttributes.LAST_POSITION_ALLOWED)) {
-      return getStyle().get(StyleAttributes.LAST_POSITION_ALLOWED);
-    }
-    return !getStyle().get(StyleAttributes.PUNCTUATION_RIGHT);
+    return StyleAttributesUtil.isLastPositionAllowed(getStyle());
   }
 
   public int getCaretPosition() {
@@ -565,7 +569,7 @@ public abstract class EditorCell_Label extends EditorCell_Basic implements jetbr
       changeText(newText);
       setCaretPositionIfPossible(startSelection + 1);
       myTextLine.resetSelection();
-      editor.resetLastCaretX();
+      fireSelectionChanged();
       ensureCaretVisible();
       return true;
     }
@@ -580,10 +584,17 @@ public abstract class EditorCell_Label extends EditorCell_Basic implements jetbr
     return label.isEditable() && label.isSelectable();
   }
 
-  private void deleteIfPossible() {
+  private void deleteIfPossible(CellActionType actionType) {
+    assert CellActionType.DELETE == actionType || CellActionType.BACKSPACE == actionType;
     if ("".equals(getText()) && isTopCell()) {
       if (getStyle().get(StyleAttributes.AUTO_DELETABLE)) {
-        getSNode().delete();
+        // TODO: just use delte action (do not call getSNode().delete()) in the end if acton was not found or is not applicable
+        CellAction deleteAction = getEditorComponent().getActionHandler().getApplicableCellAction(this, actionType);
+        if (deleteAction != null && deleteAction.canExecute(getContext())) {
+          deleteAction.execute(getContext());
+        } else {
+          getSNode().delete();
+        }
       }
     }
   }
@@ -620,7 +631,7 @@ public abstract class EditorCell_Label extends EditorCell_Basic implements jetbr
     int endSel = myTextLine.getEndTextSelectionPosition();
     changeText(myText.substring(0, stSel) + myText.substring(endSel));
     myTextLine.setCaretPosition(stSel);
-    editor.resetLastCaretX();
+    fireSelectionChanged();
     ensureCaretVisible();
   }
 
@@ -636,13 +647,13 @@ public abstract class EditorCell_Label extends EditorCell_Basic implements jetbr
 
     SNode node = getSNode();
     if (node == null) return;
-    if (CommandProcessor.getInstance().getCurrentCommand() == null)return;
+    if (CommandProcessor.getInstance().getCurrentCommand() == null) return;
     if (EqualUtil.equals(oldText, text)) return;
     if (isValidText(text)) return;
 
     UndoHelper.getInstance().addUndoableAction(new MySNodeUndoableAction(node, cellInfo, editor, oldText, text));
 
-    if (node.getModel()==null) return;
+    if (node.getModel() == null) return;
 
     MPSNodesVirtualFileSystem.getInstance().getFileFor(node.getContainingRoot()).setModificationStamp(LocalTimeCounter.currentTime());
   }
@@ -780,6 +791,10 @@ public abstract class EditorCell_Label extends EditorCell_Basic implements jetbr
     return myTextLine.getTextAfterCaret();
   }
 
+  private void fireSelectionChanged() {
+    getEditorComponent().getSelectionManager().setSelection(getEditorComponent().getSelectionManager().getSelection());
+  }
+
   private static class MySNodeUndoableAction extends SNodeUndoableAction {
     private final CellInfo myCellInfo;
     private final WeakReference<EditorComponent> myEditor;
@@ -833,7 +848,7 @@ public abstract class EditorCell_Label extends EditorCell_Basic implements jetbr
     @Override
     public void execute(EditorContext context) {
       setCaretPosition(getCaretPosition() - 1, myWithSelection);
-      ((EditorComponent) context.getEditorComponent()).resetLastCaretX();
+      fireSelectionChanged();
       ensureCaretVisible();
     }
   }
@@ -853,7 +868,7 @@ public abstract class EditorCell_Label extends EditorCell_Basic implements jetbr
     @Override
     public void execute(EditorContext context) {
       setCaretPosition(getCaretPosition() + 1, myWithSelection);
-      ((EditorComponent) context.getEditorComponent()).resetLastCaretX();
+      fireSelectionChanged();
       ensureCaretVisible();
     }
   }
@@ -867,7 +882,7 @@ public abstract class EditorCell_Label extends EditorCell_Basic implements jetbr
     @Override
     public void execute(EditorContext context) {
       setCaretPosition(0, true);
-      ((EditorComponent) context.getEditorComponent()).resetLastCaretX();
+      fireSelectionChanged();
       ensureCaretVisible();
     }
   }
@@ -881,7 +896,7 @@ public abstract class EditorCell_Label extends EditorCell_Basic implements jetbr
     @Override
     public void execute(EditorContext context) {
       setCaretPosition(getText().length(), true);
-      ((EditorComponent) context.getEditorComponent()).resetLastCaretX();
+      fireSelectionChanged();
       ensureCaretVisible();
     }
   }
@@ -968,7 +983,7 @@ public abstract class EditorCell_Label extends EditorCell_Basic implements jetbr
       EditorCell_Label cell = (EditorCell_Label) context.getSelectedCell();
       final String s = TextPasteUtil.getStringFromClipboard();
       cell.insertText(NameUtil.escapeInvisibleCharacters(s));
-      ((EditorComponent) context.getEditorComponent()).resetLastCaretX();
+      fireSelectionChanged();
       cell.ensureCaretVisible();
     }
   }
@@ -1013,7 +1028,7 @@ public abstract class EditorCell_Label extends EditorCell_Basic implements jetbr
       if (myActionType == CellActionType.BACKSPACE) {
         if (myTextLine.hasNonTrivialSelection()) {
           deleteSelection();
-          deleteIfPossible();
+          deleteIfPossible(myActionType);
           return true;
         }
 
@@ -1026,7 +1041,7 @@ public abstract class EditorCell_Label extends EditorCell_Basic implements jetbr
           if (!isCaretPositionAllowed(caretPosition - 1)) return false;
           setCaretPosition(caretPosition - 1);
           ensureCaretVisible();
-          deleteIfPossible();
+          deleteIfPossible(myActionType);
           return true;
         } else {
           if (myAllowErrors && canDeleteFrom(getPrevLeaf())) {
@@ -1041,7 +1056,7 @@ public abstract class EditorCell_Label extends EditorCell_Basic implements jetbr
       } else if (myActionType == CellActionType.DELETE) {
         if (myTextLine.hasNonTrivialSelection()) {
           deleteSelection();
-          deleteIfPossible();
+          deleteIfPossible(myActionType);
           return true;
         } else if (caretPosition < oldText.length()) {
           String newText = oldText.substring(0, caretPosition) + oldText.substring(caretPosition + 1);
@@ -1050,7 +1065,7 @@ public abstract class EditorCell_Label extends EditorCell_Basic implements jetbr
           }
           changeText(newText);
           ensureCaretVisible();
-          deleteIfPossible();
+          deleteIfPossible(myActionType);
           return true;
         } else {
           if (myAllowErrors && canDeleteFrom(getNextLeaf())) {
